refactor incremental pipeline, separating algorithm and pipeline code
cross_view_incremental.py split up into visualization code
(viz/incremental.py), pipeline code (cross_view_incremental.py)
, and algorithmic code (cross_view/incremental_localizaiton.py).

update default primitive match dim to 2.
