+++ meridian/pipeline/cross_view_localization.py
@@ -10,6 +10,10 @@ import cv2 as cv
 import matplotlib.pyplot as plt
 import numpy as np
 
+from meridian.cross_view.candidates import (
+    LocalizationContext,
+    build_candidates_from_match_result,
+)
 from meridian.cross_view.rpgo import (
     CrossViewRPGO,
     CrossViewRPGOResult,
@@ -37,106 +41,30 @@ logger = logging.getLogger(__name__)
 # ---------------------------------------------------------------------------
 # Candidate construction helper
 # ---------------------------------------------------------------------------
+#
+# `build_candidates_from_match_result` now lives in
+# `meridian.cross_view.candidates` (imported above) so the algorithm layer does
+# not depend on pipeline-layer data types. `context_from_data` bridges the
+# pipeline's `CrossViewLocalizationData` to the algorithm's `LocalizationContext`.
 
 
-def build_candidates_from_match_result(
-    match_result: CrossViewMatchResult,
-    ground_submaps: Dict[str, Submap],
-    aerial_submaps: Dict[str, Submap],
-    data: CrossViewLocalizationData,
-    min_assoc: int,
-) -> List[dict]:
-    """Build candidate list from in-memory CrossViewMatchResult."""
-    first_aerial_key = next(iter(aerial_submaps))
-    pose_flu = aerial_submaps[first_aerial_key].pose
-
-    candidates = []
-
-    for ground_key, results_matrix in match_result.results.items():
-        if ground_key not in ground_submaps:
-            continue
-        ground_submap = ground_submaps[ground_key]
-        ground_camera_pose = ground_submap.metadata["camera_pose"]
-
-        for idx in np.ndindex(results_matrix.shape):
-            cell = results_matrix[idx]
-            hypotheses = cell if isinstance(cell, list) else [cell]
-            for result in hypotheses:
-                if result.num_associations < min_assoc:
-                    continue
-                T_i_j_hat = result.T_i_j_hat
-                if np.any(np.isnan(T_i_j_hat)):
-                    continue
-
-                if data.T_camera_flu is not None:
-                    T_odom_ground = ground_camera_pose @ data.T_camera_flu
-                else:
-                    T_odom_ground = ground_camera_pose
-
-                T_full = pose_flu @ T_i_j_hat
-
-                if np.linalg.det(T_full[:2, :2]) < 0:
-                    logger.debug(
-                        "Rejecting candidate with reflected rotation "
-                        f"(det(R_2x2)={np.linalg.det(T_full[:2, :2]):.4f})"
-                    )
-                    continue
-
-                yaw = np.arctan2(T_full[1, 0], T_full[0, 0])
-
-                pixel_len_m = data.aerial_img_scale
-                body_col = T_i_j_hat[0, 3] / pixel_len_m
-                body_row = T_i_j_hat[1, 3] / pixel_len_m
-                if (
-                    data.geotiff_transform is not None
-                    and data.native_crs is not None
-                    and data.utm_crs is not None
-                    and data.native_crs != data.utm_crs
-                ):
-                    utm_x, utm_y = data.aerial_pixel_to_utm(body_col, body_row)
-                else:
-                    utm_x = T_full[0, 3]
-                    utm_y = T_full[1, 3]
-
-                T_utm_body_se2 = se2_from_xytheta(utm_x, utm_y, yaw)
-
-                T_utm_odom_4x4 = se2_to_se3(T_utm_body_se2) @ np.linalg.inv(
-                    T_odom_ground
-                )
-                T_utm_odom_se2 = se3_to_se2(T_utm_odom_4x4)
-
-                T_utm_odom_gt_se2 = None
-                if data.gt_pose_data is not None:
-                    try:
-                        gt_pose = data.gt_pose_data.pose(ground_submap.time)
-                        if data.T_camera_flu is not None:
-                            gt_body = gt_pose @ data.T_camera_flu
-                        else:
-                            gt_body = gt_pose
-                        T_utm_odom_gt = gt_body @ np.linalg.inv(T_odom_ground)
-                        T_utm_odom_gt_se2 = se3_to_se2(T_utm_odom_gt)
-                    except Exception:
-                        pass
-
-                candidates.append(
-                    {
-                        "T_utm_odom_se2": T_utm_odom_se2,
-                        "T_utm_body_se2": T_utm_body_se2,
-                        "T_i_j_hat": T_i_j_hat,
-                        "T_i_j": result.T_i_j,
-                        "aerial_pose": pose_flu,
-                        "ground_camera_pose": ground_camera_pose,
-                        "T_odom_ground": T_odom_ground,
-                        "ground_key": ground_key,
-                        "aerial_key": f"{idx[0]}_{idx[1]}",
-                        "num_associations": result.num_associations,
-                        "ground_submap_time": ground_submap.time,
-                        "T_utm_odom_gt_se2": T_utm_odom_gt_se2,
-                        "count": getattr(result, "count", 1),
-                    }
-                )
-
-    return candidates
+def context_from_data(data: CrossViewLocalizationData) -> LocalizationContext:
+    """Build a `LocalizationContext` from pipeline `CrossViewLocalizationData`."""
+    needs_crs_reprojection = (
+        data.geotiff_transform is not None
+        and data.native_crs is not None
+        and data.utm_crs is not None
+        and data.native_crs != data.utm_crs
+    )
+    return LocalizationContext(
+        T_camera_flu=data.T_camera_flu,
+        aerial_img_scale=data.aerial_img_scale,
+        local_to_pixel_fn=(
+            data.aerial_local_to_pixel if data.geotiff_transform is not None else None
+        ),
+        pixel_to_utm_fn=data.aerial_pixel_to_utm if needs_crs_reprojection else None,
+        gt_pose_data=data.gt_pose_data,
+    )
 
 
 def _count_submaps_with_below_threshold_match(
@@ -569,7 +497,11 @@ class CrossViewLocalization:
             else self.rpgo_params.min_num_associations
         )
         return build_candidates_from_match_result(
-            match_result, ground_submaps, aerial_submaps, data, min_assoc
+            match_result,
+            ground_submaps,
+            aerial_submaps,
+            context_from_data(data),
+            min_assoc,
         )
 
     # ------------------------------------------------------------------
